Gather stats.

+++ NetProxy.Service/Proxy/NpProxyEndpointStatistics.cs
@@ -3,7 +3,7 @@
     public class NpProxyEndpointStatistics
     {
         public ulong TotalConnections = 0;
-        public ulong CurrentConnections = 0;
+        public int CurrentConnections = 0;
         public ulong BytesWritten = 0;
         public ulong BytesRead = 0;
     }
